backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
openai==1.12.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
httpx==0.26.0
websockets==12.0
numpy==1.26.3
torch==2.1.2
